Web UI & Backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
snowflake-connector-python==3.0.4
tabulate==0.9.0
PyGithub==1.58.2
langchain==0.0.312
langchain-core==0.1.4
huggingface_hub==0.16.4
langgraph==0.0.43
langgraph-swarm==0.0.6
requests==2.31.0
